services/ingestion/requirements.txt
fastapi
uvicorn[standard]
google-genai
qdrant-client
python-dotenv
python-docx
pypdf
pdf2image
pytesseract
Pillow
numpy
python-multipart
sentence-transformers==2.2.2
torch>=2.0
